Valerian's suggestions
