+++ webapp/src/lib/dialog-transaction.scss
